image/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
pydantic>=2.7.0
pydantic-settings>=2.2.0
openai>=1.30.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
pdfplumber>=0.11.0
chromadb==1.3.5
numpy==1.26.4
pandas>=2.2.0
scikit-learn>=1.4.0
tqdm>=4.66.0
langchain_text_splitters
streamlit
pyarrow==14.0.2